+++ libs/wire-api/src/Wire/API/Routes/Public/Brig.hs
@@ -44,6 +44,7 @@ import Wire.API.ErrorDescription
   )
 import Wire.API.Routes.MultiVerb
 import Wire.API.Routes.Public (EmptyResult, ZConn, ZUser)
+import Wire.API.Routes.QualifiedCapture
 import Wire.API.User
 import Wire.API.User.Client
 import Wire.API.User.Client.Prekey
@@ -64,6 +65,8 @@ type GetUserVerb =
 
 type CaptureUserId name = Capture' '[Description "User Id"] name UserId
 
+type QualifiedCaptureUserId name = QualifiedCapture' '[Description "User Id"] name UserId
+
 type CaptureClientId name = Capture' '[Description "ClientId"] name ClientId
 
 type NewClientResponse = Headers '[Header "Location" ClientId] Client
@@ -83,8 +86,7 @@ data Api routes = Api
         :- Summary "Get a user by Domain and UserId"
         :> ZUser
         :> "users"
-        :> Capture "domain" Domain
-        :> CaptureUserId "uid"
+        :> QualifiedCaptureUserId "uid"
         :> GetUserVerb,
     getSelf ::
       routes
@@ -112,8 +114,7 @@ data Api routes = Api
         :> ZUser
         :> "users"
         :> "by-handle"
-        :> Capture "domain" Domain
-        :> Capture' '[Description "The user handle"] "handle" Handle
+        :> QualifiedCapture' '[Description "The user handle"] "handle" Handle
         :> MultiVerb
              'GET
              '[JSON]
@@ -151,8 +152,7 @@ data Api routes = Api
       routes
         :- Summary "Get all of a user's clients."
         :> "users"
-        :> Capture "domain" Domain
-        :> CaptureUserId "uid"
+        :> QualifiedCaptureUserId "uid"
         :> "clients"
         :> Get '[JSON] [PubClient],
     getUserClientUnqualified ::
@@ -167,8 +167,7 @@ data Api routes = Api
       routes
         :- Summary "Get a specific client of a user."
         :> "users"
-        :> Capture "domain" Domain
-        :> CaptureUserId "uid"
+        :> QualifiedCaptureUserId "uid"
         :> "clients"
         :> CaptureClientId "client"
         :> Get '[JSON] PubClient,
@@ -203,8 +202,7 @@ data Api routes = Api
         :- Summary "Get a prekey for a specific client of a user."
         :> ZUser
         :> "users"
-        :> Capture "domain" Domain
-        :> CaptureUserId "uid"
+        :> QualifiedCaptureUserId "uid"
         :> "prekeys"
         :> CaptureClientId "client"
         :> Get '[JSON] ClientPrekey,
@@ -221,8 +219,7 @@ data Api routes = Api
         :- Summary "Get a prekey for each client of a user."
         :> ZUser
         :> "users"
-        :> Capture "domain" Domain
-        :> CaptureUserId "uid"
+        :> QualifiedCaptureUserId "uid"
         :> "prekeys"
         :> Get '[JSON] PrekeyBundle,
     getMultiUserPrekeyBundleUnqualified ::